win back a tiny bit of performance for 64B cachelines
this is kind of punting on structure layout because
alignment attributes are a nuisance cross-platform.
this seems to speed up load of 'capacity' because
ping-pong on buffer markers seems to be adding some
latency to relaxed loads of 'capacity' likely due
to invalidating its cache line.

broke this in 3d604c286be1d98d594bd99807343e947e5e45f0
because forget about it when removing a struct member.