fix non-critical bounds bug in test precomputation
the loop that precomputes a checksum of integers limits
precomputation to the buffer size to limit the amount of
array that is touched for each read or write call, so the
second bounds is only there to limit cache eviction. fix
a bug in second bounds. we forgot to shift bufsize >> 2.